inference/requirements.txt
flash-attn==2.5.8
numpy==1.26.4
spconv-cu117==2.3.6